CU-86b2czk01 Refactor realtabformer
