+++ pyproject.toml
@@ -64,6 +64,7 @@ sdgym = { main = 'sdgym.cli.__main__:main' }
 
 [project.optional-dependencies]
 dask = ['dask', 'distributed']
+realtabformer = ['realtabformer>=0.2.1']
 test = [
     'pytest>=6.2.5',
     'pytest-cov>=2.6.0',
