+++ sdgym/__init__.py
@@ -15,7 +15,7 @@ import logging
 from sdgym.benchmark import benchmark_single_table
 from sdgym.cli.collect import collect_results
 from sdgym.cli.summary import make_summary_spreadsheet
-from sdgym.sdgym_datasets import get_available_datasets, load_dataset
+from sdgym.datasets import get_available_datasets, load_dataset
 from sdgym.synthesizers import create_sdv_synthesizer_variant, create_single_table_synthesizer
 
 # Clear the logging wrongfully configured by tensorflow/absl
