+++ sdgym/cli/utils.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import pandas as pd
 import tqdm
 
-from sdgym.sdgym_datasets import DATASETS_PATH, load_dataset
+from sdgym.datasets import DATASETS_PATH, load_dataset
 from sdgym.s3 import get_s3_client, is_s3_path, parse_s3_path
